auth-pam: Immediately report interactive instructions to clients
SSH keyboard-interactive authentication method supports instructions but
sshd didn't show them until an user prompt was requested.

This is quite inconvenient for various PAM modules that need to notify
an user without requiring for their explicit input.

So, properly implement RFC4256 making instructions to be shown to users
when they are requested from PAM.

A new KbdInteractiveLegacyInstructions option is added to the server to
disable this new behavior.

Even though sshd is not supporting ssh1 anymore, this may still be
convenient in order to support some legacy client that could be broken
by this change.

Closes: https://bugzilla.mindrot.org/show_bug.cgi?id=2876