auth-pam: Add pam-legacy-instructions device type
As per previous changes, old clients may not be able to use interactive
keyboard authentication.

In order to preserve backward compatibility for such old clients without
requiring changes on their side, we add a new interactive device support
to sshd, so that old clients can require it via:
 -o KbdInteractiveDevices=pam-legacy-instructions

And this option should be supported by most clients out there.

The device we create is actually re-using most of the PAM device code a
part for the initialization bits.